updating hello.sh in examples again
adding back in the backslash *because* while normally you wouldn't need to add the backslash for the single quote within double quoted string there is apparently some other code within my version of bash or desk (which I have not tried to track down but probably one of those sed lines) that parses the strings from the deskfiles and ultimately does some replacements causing this line to become: 

alias howdy='echo howdy y\'\''all'

This replacement fixes the problems with the single/double quotes and means that in the deskfiles we need to use the backslash character before the single quote within the double quotes values. All other possibilities I've tried have not worked and it seems to be due to these character replacements methods happening somewhere between saving the file, running desk, and trying the alias. 

Also, when removing the backslash the CI build breaks but with the backslash it seems to compile correctly... at least I think it did last week. With the backslash it "works for me" everytime. 

If it doesn't work for you please let me know and I'll figure out how to close this PR without further commits or bother. I do like the software you've created and will keep it configured like this on my clone. 

Thanks y'all.