Adding test scenarios for displaying exported variables
This commit includes two file updates for testing the new feature
as requested in jamesob/desk#56. The example hello deskfile now has
a variable set called MyName with a value "James" and the runtests
shell script includes a test to echo the value of that variable.

This test proves that the variable is exported as expected from the
deskfile and is retrievable as expected.

Still #todo includes a test that lists the description of the deskfile
and looks for that variable to be listed along with the aliases and
functions.